lesson 28.10.2024 update

+++ src/service/MainServiceImpl.java
@@ -1,79 +1,106 @@
 package service;
 
 import model.Car;
+import model.Role;
 import model.User;
 import repository.CarRepository;
 import repository.UserRepository;
 import utils.MyList;
 
-public class MainServiceImpl implements MainService{
+/**
+ * @author Sergey Bugaenko
+ * {@code @date} 25.10.2024
+ */
 
-    // Поля
+public class MainServiceImpl implements MainService {
 
     private final CarRepository repositoryCar;
     private final UserRepository repositoryUser;
     private User activeUser;
 
-    // Конструктор
-
-    public MainServiceImpl(CarRepository repositoryCar, UserRepository repositoryUser) {
-        this.repositoryCar = repositoryCar;
+    public MainServiceImpl(CarRepository carRepository, UserRepository repositoryUser) {
+        this.repositoryCar = carRepository;
         this.repositoryUser = repositoryUser;
     }
 
-    // Методы
 
-    // Добавляет автомобиль в список
     @Override
     public void addCar(String model, int year, double price) {
+        // Проверка роли текущего пользователя
+        if (activeUser == null || activeUser.getRole() != Role.ADMIN) {
+            System.out.println("Добавление нового автомобиля доступно только Администраторам");
+            return;
+        }
+
+        // Проверка входящих значений и добавление автомобиля в БД
 
     }
 
-    // Возвращает текущий список автомобилей
     @Override
     public MyList<Car> getAllCars() {
         return repositoryCar.getAllCars();
     }
 
-    // Возвращает список автомобилей по модели
     @Override
     public MyList<Car> getCarsByModel(String model) {
         return null;
     }
 
-    // Возвращает список свободных автомобилей
     @Override
     public MyList<Car> getFreeCars() {
         return null;
     }
 
-    // Обновляет цену автомобиля
     @Override
     public boolean updateCarPrice(int id, double price) {
-        Car car = repositoryCar.getByID(id);
+        // Добавим проверку роли пользователя
+        if (activeUser == null || activeUser.getRole() != Role.ADMIN) {
+            System.out.println("Изменять цену может только администратор");
+            return false;
+        }
+
+        Car car = repositoryCar.getById(id);
         if (car == null || price < 0) {
             return false;
         }
+
         car.setPrice(price);
         return true;
     }
 
-    // Осуществляет взятие автомобиля в аренду и возвращает статус успеха операции
     @Override
     public boolean takeCar(int id) {
-        Car car = repositoryCar.getByID(id);
+        // Проверка, что пользователь авторизирован в системе
+        if (activeUser == null) {
+            System.out.println("Выполните вход в систему");
+            return false;
+        }
+
+        if (activeUser.getRole() == Role.BLOCKED) {
+            System.out.println("Ваш аккаунт заблокирован! Обратитесь в службу поддержки");
+            return false;
+        }
+
+        Car car = repositoryCar.getById(id);
+        // Todo
         /*
-        1. Проверить, что автомобиль найден и свободен. Если нет - закончить метод;
-        2. Если да: пометить автомобиль как занятый;
-        3. Добавить автомобиль в список текущего пользователя
+        1. Проверить, что машина найден и свободна. Если нет - закончить метод
+        2. Пометить машину, как занятую
+        3. Добавить машину в список машин текущего пользователя
          */
+
         return false;
     }
 
-    // Удаляет автомобиль из хранилища
     @Override
     public Car deleteCar(int id) {
-        Car car = repositoryCar.getByID(id);
+        // Проверка пользователя
+        if (activeUser == null || activeUser.getRole() != Role.ADMIN) {
+            System.out.println("Доступно только Администратору");
+            return null;
+        }
+        
+        Car car = repositoryCar.getById(id);
         if (car == null) return null;
 
         repositoryCar.deleteCar(car);
@@ -85,7 +112,7 @@ public class MainServiceImpl implements MainService{
         // Добавить валидацию email и password
         // Если не прошли валидацию - закончить работу метода, вернуть null
 
-        if(repositoryUser.isEmailExists(email)) {
+        if (repositoryUser.isEmailExists(email)) {
             System.out.println("Email already exists");
             return null;
         }
@@ -97,13 +124,12 @@ public class MainServiceImpl implements MainService{
 
     @Override
     public boolean loginUser(String email, String password) {
-
         /*
-        1. Нужно найти в БД пользователя с таким email
+        1. Нужно пользователя в БД с таким email
         2. Если пользователя не существует - отказать
-        3. Проверить, совпадает ли пароль пользователя в БД с паролем, который пришёл в методе
-        4. Если не совпадает, отказать;
-        5. Залогинить пользователя - пометить как активного пользователя системы
+        3. Проверить, совпадает ли пароль у пользователя в БД с паролем, который пришел в методе
+        4. Если не совпадает - отказать
+        5. Залогинить пользователя - пометить как активный пользователь системы
          */
 
         User user = repositoryUser.getUserByEmail(email);
@@ -112,7 +138,7 @@ public class MainServiceImpl implements MainService{
             return false;
         }
 
-        if(!user.getPassword().equals(password)) {
+        if (!user.getPassword().equals(password)) {
             System.out.println("Invalid email or password");
             return false;
         }
